cleans up some code
