+++ deployments/Dockerfile
@@ -1,3 +1,4 @@
 FROM rust
 
-RUN cargo install systemfd cargo-watch
+RUN cargo install systemfd cargo-watch diesel_cli
+RUN diesel run migration
