+++ src/api/mod.rs
@@ -1,0 +1,2 @@
+pub mod routes;
+pub mod index;
